Further split graphiql setup appart
