+++ packages/cli/src/commands/setup/graphiql/__tests__/graphiqlHandler.test.js
@@ -21,7 +21,8 @@ jest.mock('listr2')
 import chalk from 'chalk'
 import { Listr } from 'listr2'
 
-import * as graphiql from '../graphiql'
+import * as graphiqlHandler from '../graphiqlHandler'
+import * as graphiqlHelpers from '../graphiqlHelpers'
 
 describe('Graphiql generator tests', () => {
   const processExitSpy = jest
@@ -42,8 +43,8 @@ describe('Graphiql generator tests', () => {
   })
 
   it('throws an error if source path does not exist when viewing headers', async () => {
-    jest.spyOn(graphiql, 'getOutputPath').mockImplementation(() => '')
-    await graphiql.handler({ view: true, provider: 'dbAuth' })
+    jest.spyOn(graphiqlHelpers, 'getOutputPath').mockImplementation(() => '')
+    await graphiqlHandler.handler({ view: true, provider: 'dbAuth' })
     expect(console.error).toHaveBeenCalledWith(
       chalk.bold.red(
         'Must run yarn rw setup graphiql <provider> to generate headers before viewing'
@@ -54,7 +55,7 @@ describe('Graphiql generator tests', () => {
 
   it('throws an error if auth provider is dbAuth and no user id is provided', () => {
     try {
-      graphiql.generatePayload('dbAuth')
+      graphiqlHelpers.generatePayload('dbAuth')
     } catch (e) {
       expect(e.message).toBe('Require an unique id to generate session cookie')
     }
@@ -63,7 +64,7 @@ describe('Graphiql generator tests', () => {
   it('throws an error if auth provider is dbAuth and no supabase env is set', () => {
     process.env.SESSION_SECRET = null
     try {
-      graphiql.generatePayload('dbAuth', 'user-id-123')
+      graphiqlHelpers.generatePayload('dbAuth', 'user-id-123')
     } catch (e) {
       expect(e.message).toBe(
         'dbAuth requires a SESSION_SECRET environment variable that is used to encrypt session cookies. Use `yarn rw g secret` to create one, then add to your `.env` file. DO NOT check this variable in your version control system!!'
@@ -74,7 +75,7 @@ describe('Graphiql generator tests', () => {
   it('returns a payload if a token is provided', async () => {
     const provider = 'supabase'
     const token = 'mock-token'
-    const response = graphiql.generatePayload(provider, null, token)
+    const response = graphiqlHelpers.generatePayload(provider, null, token)
     expect(response).toEqual({
       'auth-provider': provider,
       authorization: `Bearer ${token}`,
